requirements.txt
absl-py==1.4.0
affine==2.3.1
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
docker-pycreds==0.4.0
earthengine-api==0.1.336
efficientnet==1.0.0
flatbuffers==23.1.4
fonttools==4.38.0
future==0.18.3
gast==0.5.3
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
h5py==3.7.0
httplib2==0.21.0
idna==3.4
image-classifiers==1.0.0
imageio==2.24.0
importlib-metadata==6.0.0
joblib==1.2.0
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-unet-collection==0.0.7b0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
pathtools==0.1.2
Pillow==9.4.0
plotly==5.12.0
protobuf==3.20.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
rasterio==1.2.10
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
segmentation-models==1.0.1
sentry-sdk==1.13.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
snuggs==1.4.7
tenacity==8.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2021.11.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
wandb==0.13.9
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
